requirements.txt
appdirs==1.4.3
awscli==1.18.2
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.12.2
botocore==1.15.2
Brotli==1.0.7
certifi==2019.6.16
cffi==1.14.0
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.5
django-admin-sortable2==0.7.5
django-bootstrap4==2.1.1
django-ckeditor==5.7.1
django-compression-middleware==0.4.1
django-cors-headers==3.2.0
django-filter==2.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-richtextfield==1.4.0
django-storages==1.9.1
django-tables2==2.3.1
django-taggit==1.1.0
django-tinymce==2.8.0
djangorestframework==3.11.0
docutils==0.14
filelock==3.0.12
gunicorn==20.0.4
jmespath==0.9.4
leather==0.3.3
Markdown==3.1.1
Pillow==7.0.0
pipenv==2018.11.26
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.2
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==20.0.10
virtualenv-clone==0.5.3
whitenoise==4.1.3
zstandard==0.13.0

wagtail~=2.10